requirements.txt
python-telegram-bot
feedparser
py-postgresql